Update to Ruff 0.16
https://github.com/astral-sh/ruff/blob/main/CHANGELOG.md

Changes required:

```
[ruff:check] $ ruff check .
EXE001 Shebang is present but file is not executable
 --> sox/annotate_waveform.py:1:1
  |
1 | #!/usr/bin/env -S uv run --script
  | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
2 | #
3 | # /// script
  |

RUF046 Value being cast to `int` is already an integer
  --> sox/annotate_waveform.py:58:15
   |
57 | def format_time(seconds: float) -> str:
58 |     rounded = int(round(seconds))
   |               ^^^^^^^^^^^^^^^^^^^
59 |     hours = rounded // 3600
60 |     minutes = (rounded % 3600) // 60
   |
help: Remove unnecessary `int` call

Found 2 errors.
No fixes available (1 hidden fix can be enabled with the `--unsafe-fixes` option).
```

+++ sox/annotate_waveform.py
@@ -55,7 +55,7 @@ def text_size(draw: ImageDraw.ImageDraw, text: str) -> tuple[int, int]:
 
 
 def format_time(seconds: float) -> str:
-    rounded = int(round(seconds))
+    rounded = round(seconds)
     hours = rounded // 3600
     minutes = (rounded % 3600) // 60
     secs = rounded % 60
